refactor: 💡 update import path of theme const
